+++ packages/granularity/src/components/GrEmptyState/config.ts
@@ -1,6 +1,9 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grEmptyStateSafelist } from './safelist'
+
 export const grEmptyStateConfig = defineGranularComponent(import.meta.url, {
   name: 'GrEmptyState',
+  safelist: grEmptyStateSafelist,
   dependencies: ['GrIcon'],
 })