+++ packages/granularity/src/components/GrEmptyState/index.ts
@@ -2,3 +2,11 @@ export { default } from './GrEmptyState.vue'
 export { default as GrEmptyState } from './GrEmptyState.vue'
 export type { GrEmptyStateProps } from './GrEmptyState.vue'
 export { grEmptyStateConfig } from './config'
+export type { GrEmptyStateConfigurableProps } from './defaults'
+export { GR_EMPTY_STATE_HEADING_LEVELS, GR_EMPTY_STATE_VARIANTS } from './grEmptyStateStyles'
+export type {
+  GrEmptyStateHeadingLevel,
+  GrEmptyStateSize,
+  GrEmptyStateVariant,
+} from './grEmptyStateStyles'
+export { grEmptyStateSafelist } from './safelist'